pyproject.toml
[project]
name = "digital-pensieve"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "mcp[cli]>=1.7.1",
    "sentence-transformers>=4.1.0",
    "sounddevice>=0.5.1",
]
